Seperate user lock info logic into own component + Show env and deployment details in tooltip
