+++ client/src/app/components/user-lock-info/user-lock-info.component.html
@@ -1,0 +1,19 @@
+@if (latestLock(); as lock) {
+  <a routerLink="environment/list" class="text-sm text-gray-500 flex flex-col items-center gap-1 mb-5" [pTooltip]="tooltipContent">
+    <i-tabler name="lock"></i-tabler>
+    <span class="w-20 text-center">{{ timeSinceLocked() }}</span>
+  </a>
+  <ng-template #tooltipContent>
+    <div class="flex flex-col">
+      <span class="text-xs text-gray-200 uppercase tracking-tighter font-bold">Locked Environment</span>
+      <span class="text-sm">{{ lock.environment?.name }}</span>
+      @if (lock.environment?.latestDeployment?.ref) {
+        <div class="border-b border-gray-500 my-2"></div>
+        <div class="flex gap-1 items-center">
+          <span class="text-sm truncate">{{ lock.environment?.latestDeployment?.ref }}</span>
+          <i-tabler style="height: 16px; width: 16px" name="git-branch" class="text-gray-200 flex-shrink-0" />
+        </div>
+      }
+    </div>
+  </ng-template>
+}
